+++ taskgraph/Task.py
@@ -84,7 +84,7 @@ class TaskGraph(object):
 
     def __init__(
             self, taskgraph_cache_dir_path, n_workers,
-            reporting_interval=None):
+            reporting_interval=None, delayed_start=False):
         """Create a task graph.
 
         Creates an object for building task graphs, executing them,
@@ -101,11 +101,24 @@ class TaskGraph(object):
                 `add_task` will be a blocking call.
             reporting_interval (scalar): if not None, report status of task
                 graph every `reporting_interval` seconds.
+            delayed_start (bool): if true, taskgraph does not start executing
+                tasks as `add_task` is called. Instead no execution occurs
+                until `join` is invoked. A value of `True` is incompatible
+                with `n_workers` < 0 and will raise a ValueError on
+                construction.
 
         """
+        if delayed_start and n_workers < 0:
+            raise ValueError(
+                "`n_workers` cannot be set single process mode while "
+                "`delayed_start` is enabled.")
         # the work queue is the feeder to active worker threads
         self.taskgraph_cache_dir_path = taskgraph_cache_dir_path
         self.n_workers = n_workers
+        self.taskgraph_started_event = threading.Event()
+        if not delayed_start:
+            # not a delayed start, so clear the event immediately
+            self.taskgraph_started_event.set()
 
         # keep track if the task graph has been forcibly terminated
         self.terminated = False
@@ -157,29 +170,43 @@ class TaskGraph(object):
         self.work_queue = queue.Queue()
         self.worker_semaphore = threading.Semaphore(max(1, n_workers))
         # launch threads to manage the workers
+        self.worker_thread_list = []
         for thread_id in xrange(max(1, n_workers)):
             worker_thread = threading.Thread(
                 target=self._task_worker,
                 name='taskgraph_worker_thread_%d' % thread_id)
             worker_thread.daemon = True
             worker_thread.start()
+            self.worker_thread_list.append(worker_thread)
 
         # tasks that get passed add_task get put in this queue for scheduling
         self.waiting_task_queue = queue.Queue()
-        waiting_task_scheduler = threading.Thread(
+        self.waiting_task_scheduler = threading.Thread(
             target=self._process_waiting_tasks,
             name='_waiting_task_scheduler')
-        waiting_task_scheduler.daemon = True
-        waiting_task_scheduler.start()
+        self.waiting_task_scheduler.daemon = True
+        self.waiting_task_scheduler.start()
 
         # tasks in the work ready queue have dependencies satisfied but need
         # priority scheduling
         self.work_ready_queue = queue.Queue()
-        priority_task_scheduler = threading.Thread(
+        self.priority_task_scheduler = threading.Thread(
             target=self._schedule_priority_tasks,
             name='_priority_task_scheduler')
-        priority_task_scheduler.daemon = True
-        priority_task_scheduler.start()
+        self.priority_task_scheduler.daemon = True
+        self.priority_task_scheduler.start()
+
+    def __del__(self):
+        """Ensure all threads have been joined for cleanup."""
+        if self.n_workers >= 0:
+            # there's only something to clean up if there's a worker
+            LOGGER.error("joining all the workers")
+            self.priority_task_scheduler.join()
+            self.waiting_task_scheduler.start()
+            for worker_thread in self.worker_thread_list:
+                worker_thread.join()
+            if self.worker_pool:
+                self.worker_pool.join()
 
     def _task_worker(self):
         """Execute and manage Task objects."""
@@ -234,7 +261,8 @@ class TaskGraph(object):
             priority (numeric): the priority of a task is considered when
                 there is more than one task whose dependencies have been
                 met and are ready for scheduling. Tasks are inserted into the
-                work queue in order of decreasing priority. This value can be
+                work queue in order of decreasing priority value
+                (priority 10 is higher than priority 1). This value can be
                 positive, negative, and/or floating point.
 
         Returns:
@@ -266,7 +294,8 @@ class TaskGraph(object):
             new_task = Task(
                 task_name, func, args, kwargs, target_path_list,
                 ignore_path_list, dependent_task_list, ignore_directories,
-                self.worker_pool, self.taskgraph_cache_dir_path, priority)
+                self.worker_pool, self.taskgraph_cache_dir_path, priority,
+                self.taskgraph_started_event)
             task_hash = new_task.task_hash
 
             # it may be this task was already created in an earlier call,
@@ -336,7 +365,8 @@ class TaskGraph(object):
         while not stopped:
             while True:
                 try:
-                    # only block if the priority queque is empty
+                    self.taskgraph_started_event.wait()
+                    # only block if the priority queue is empty
                     task = self.work_ready_queue.get(not priority_queue)
                     if task == 'STOP':
                         # encounter STOP so break and don't get more elements
@@ -378,8 +408,13 @@ class TaskGraph(object):
         task_dependent_map = collections.defaultdict(set)
         dependent_task_map = collections.defaultdict(set)
         completed_tasks = set()
+
+        # it's possible the taskgraph is in delayed execution mode, don't
+        # attempt to schedule until signaled to do so
+        self.taskgraph_started_event.wait()
+
+        tasks_ready_to_work = set()
         for task, mode in iter(self.waiting_task_queue.get, 'STOP'):
-            tasks_ready_to_work = set()
             if mode == 'wait':
                 # see if this task's dependencies are satisfied, if so send
                 # to work.
@@ -418,10 +453,22 @@ class TaskGraph(object):
                         # work queue
                         tasks_ready_to_work.add(waiting_task)
                 del task_dependent_map[task]
-            for ready_task in sorted(
-                    tasks_ready_to_work, key=lambda x: x.priority):
-                self.work_ready_queue.put(ready_task)
-            tasks_ready_to_work = None
+            if self.waiting_task_queue.empty():
+                # this only schedules tasks if the queue is drained, fixes
+                # a race condition where a lower priority task may be
+                # immediately put to the work queue even though there are
+                # higher priority ones still waiting
+                for ready_task in sorted(
+                        tasks_ready_to_work, key=lambda x: x.priority):
+                    self.work_ready_queue.put(ready_task)
+                tasks_ready_to_work = set()
+
+        # it's possible the last element in the queue was 'STOP', this drains
+        # the `tasks_ready_to_work` set so everything gets scheduled
+        for ready_task in sorted(
+                tasks_ready_to_work, key=lambda x: x.priority):
+            self.work_ready_queue.put(ready_task)
+
         # if we got here, the waiting task queue is shut down, pass signal
         # to the lower queue
         self.work_ready_queue.put('STOP')
@@ -441,6 +488,8 @@ class TaskGraph(object):
         # if single threaded, nothing to join.
         if self.n_workers < 0:
             return True
+        # start delayed execution if necessary:
+        self.taskgraph_started_event.set()
         try:
             timedout = False
             for task in itervalues(self.task_id_map):
@@ -489,7 +538,7 @@ class Task(object):
     def __init__(
             self, task_name, func, args, kwargs, target_path_list,
             ignore_path_list, dependent_task_list, ignore_directories,
-            worker_pool, cache_dir, priority):
+            worker_pool, cache_dir, priority, taskgraph_started_event):
         """Make a Task.
 
         Parameters:
@@ -522,6 +571,8 @@ class Task(object):
                 met and are ready for scheduling. Tasks are inserted into the
                 work queue in order of decreasing priority. This value can be
                 positive, negative, and/or floating point.
+            taskgraph_started_event (Event): if set, `Task`s can execute, but
+                if not, an exception will be raised if `join` is invoked.
 
         """
         self.task_name = task_name
@@ -534,8 +585,10 @@ class Task(object):
         self.ignore_path_list = ignore_path_list
         self.ignore_directories = ignore_directories
         self.worker_pool = worker_pool
-        # invert the priority since heapq goes smallest to largest
+        # invert the priority since heapq goes smallest to largest and we
+        # want more positive priority values to be executed first.
         self.priority = -priority
+        self.taskgraph_started_event = taskgraph_started_event
 
         self.terminated = False
         self.exception_object = None
@@ -765,6 +818,10 @@ class Task(object):
 
     def join(self, timeout=None):
         """Block until task is complete, raise exception if runtime failed."""
+        if not self.taskgraph_started_event.is_set():
+            raise RuntimeError(
+                "Task joined even though taskgraph has delayed start "
+                "enabled: %s" % self)
         self._task_complete_event.wait(timeout)
         return self.is_complete()