+++ HISTORY.rst
@@ -8,6 +8,11 @@ Unreleased Changes
 * Fixing an issue where a non-JSON serializeable object would cause `add_task`
   to crash. Now TaskGraph is more tolerant of non-JSON serializeable objects
   and will log warnings when parameters cannot be serialized.
+* Added a `delayed_start` flag to TaskGraph to allow for delayed execution of
+  taskgraph tasks. If enabled on threaded or multiprocess model, calls to
+  `add_task` will not execute tasks until the `join` method is invoked on
+  `taskgraph`. This allows for finer control over execution order when tasks
+  are passed non-equivalent `priority` levels.
 
 0.5.2 (2018-06-20)
 ------------------